refactor: update commans file structure
